requirements.txt
sqlalchemy==2.0.23
alembic==1.13.1
pymysql==1.1.0
pytest==7.4.0
fastapi==0.104.1
uvicorn==0.24.0.post1
python-dotenv==1.0.0